refactor: improve next config modification logic by removing webVitalsAttribution and adding invariant checks to property removals.
