+++ src/features/ProfileEditor/index.ts
@@ -1,0 +1,2 @@
+export { default as AgentTool, type AgentToolProps } from './AgentTool';
+export { default as PluginTag, type PluginTagProps } from './PluginTag';
